added update pet with form data updated pet utils

+++ API_Test_Framework/src/test/java/com/sparta/PetApi/PetTests/DeletePetTests.java
@@ -19,7 +19,7 @@ public class DeletePetTests {
     @BeforeAll
     static void beforeAll(){
         Pet pet = PetUtils.createPetPOJO();
-        ID = pet.getId();
+        ID = (int) pet.getId();
         PetUtils.addPet(pet);
         response = PetUtils.deletePet(ID);
     }
